+++ R/RogueTaxa.R
@@ -11,7 +11,7 @@
 #' in conflict with other data \insertCite{Kearney2002}{Rogue}.
 #'
 #' These functions use heuristic methods to identify rogue taxa whose removal
-#' improves the information content of a conesnsus tree, by the definitions
+#' improves the information content of a consensus tree, by the definitions
 #' of information discussed below.
 #'
 #' @section Information criteria:
@@ -95,6 +95,8 @@ RogueTaxa <- function (trees,
                        mreOptimization = FALSE,
                        threshold = 50,
                        verbose = FALSE) {
+  p <- threshold / 100
+
   # Check format of `trees`
   if (!inherits(trees, 'multiPhylo')) {
     if (inherits(trees, 'phylo')) {
@@ -102,7 +104,7 @@ RogueTaxa <- function (trees,
                         taxNum = NA_character_,
                         taxon = NA_character_,
                         rawImprovement = NA_real_,
-                        IC = ConsensusInfo(c(trees), info = info[1]),
+                        IC = ConsensusInfo(c(trees), info = info[1], p = p),
                         stringsAsFactors = FALSE))
     }
     trees <- structure(trees, class = 'multiPhylo')
@@ -129,8 +131,8 @@ RogueTaxa <- function (trees,
   if (any(duplicated(leaves))) {
     stop("All leaves must bear unique labels.")
   }
-  trees[] <- lapply(trees, RenumberTips, trees[[1]])
-  trees[] <- lapply(trees, Preorder)
+  trees <- lapply(trees, RenumberTips, trees[[1]])
+  trees <- structure(lapply(trees, Preorder), class = 'multiPhylo')
 
   # Select and apply method
   info <- tolower(info[1])
@@ -153,11 +155,13 @@ RogueTaxa <- function (trees,
                                mreOptimization = mreOptimization,
                                threshold = threshold, verbose = verbose),
                    Roguehalla(trees, dropsetSize = dropsetSize, info = 'phylo',
-                              neverDrop = neverDrop),
+                              p = p, neverDrop = neverDrop),
                    Roguehalla(trees, dropsetSize = dropsetSize, info = 'clust',
+                              p = p, neverDrop = neverDrop),
+                   QuickRogue(trees, info = 'phylo', p = p,
                               neverDrop = neverDrop),
-                   QuickRogue(trees, info = 'phylo', neverDrop = neverDrop),
-                   QuickRogue(trees, info = 'clust', neverDrop = neverDrop)
+                   QuickRogue(trees, info = 'clust', p = p,
+                              neverDrop = neverDrop)
   )
 
   # Format return value
@@ -171,9 +175,9 @@ RogueTaxa <- function (trees,
   if (returnTree) {
     drops <- unlist(strsplit(result[-1, 'taxon'], ','))
     if (is.null(drops)) {
-      Consensus(trees, p = threshold / 100)
+      Consensus(trees, p = p)
     } else {
-      ConsensusWithout(trees, drops, p = threshold / 100)
+      ConsensusWithout(trees, drops, p = p)
     }
   } else {
     result